Fix: Check if key is already available into index
Issue #11:
When a package is store twice (via API), the storage file is created
(with a new content key), then the database insert failed and content
key in database is not updated.
After this operation, If you want to download a package, the content key
in the license file is not the content key of the package file.

Try a "Get" in index before storing file and write informations to
the database.