Refactoring of the license generation
Fix #163, FIx #165
Global refactoring of the code implementing get license, generate license, get licensed publication, generate licensed publication.
This code was difficult to maintain, full of patches.
The license table does not contain the hint, hashed passphrase and algorithm anymore. This allows for the hint and passphrase to be what the frontend decides. A script must be applied to the db, found in dbmodel, sqlite-evol-1.sql.