Readium and W3C manifest, enhancement on date-time properties
date-time properties are treated internally as time.Time, their json marshalling is tuned as per the W3C and Readium specs.
Also, a few const are transformed to CamelCase as per Go lint.
And the Readium manifest type property is either Audiobook or CreativeWork, depending the value of W3C manifest conformsTo property.